requirements.txt
flask
python-dotenv
pip
psycopg2